[#33] refactor: 리뷰 반영
Inventory 패키지 경로 수정
